Use Google Spreadsheets instead of DSM to manage color tokens
