+++ .github/CONTRIBUTING.md
@@ -135,55 +135,20 @@ Design tokens source files are in the `./tokens` directory, in YAML format.
 Adding, removing, and updating design tokens should be discussed in the issues of this project before contributing to code changes.
 
 Note: `./tokens/colors.yml` should not be manually edited, as it is automatically generated by this project.
-Colors are managed in [Invision Design System Manager](https://www.invisionapp.com/blog/announcing-invision-design-system-manager/) (see below).
+Colors are managed in a [Google Spreadsheet](https://docs.google.com/spreadsheets/d/1HzVgcxnnOCeB2zI46gcSd5fPFUJ2R03CSldx_SKQQwA/edit#gid=0) (see below).
 
 ---
 
 ## Updating color tokens
 
-Color names, values and descriptions are managed in [Invision Design System Manager](https://www.invisionapp.com/blog/announcing-invision-design-system-manager/) at this URL:
-<https://shopify.invisionapp.com/dsm/shopify/design-tokens>.
-
-### Adding colors
-
-1.  Install the Craft Plugin in [Sketch](https://www.sketchapp.com/) and log in with your Shopify account.
-2.  Once Craft for Sketch is set up, open Design System Manager, and follow these steps:
-
-    ![ ](https://user-images.githubusercontent.com/85783/38760431-0e55954e-3f30-11e8-935c-7ba35b0edc44.png)
-
-    ![ ](https://user-images.githubusercontent.com/85783/38760487-60970130-3f30-11e8-8a45-e02b7c8c4b46.png)
-
-3.  Finally, **[open a new issue][issue-template-color-updated]**
-
-### Changing a color’s name
-
-Color names are editable in <https://shopify.invisionapp.com/dsm/shopify/design-tokens> or via the Sketch Craft plugin, under the Design System Manager tab.
-
-Important considerations before changing color names: this type of update is a breaking change,
-and products consuming design tokens will have to update all occurrences of the old name in their codebases.
-
-1.  Change the name of the color
-2.  **[Open a new issue][issue-template-color-updated]**
-
-### Editing a color’s value
-
-At the time these lines were written, Invision Design System Manager doesn’t support editing colors.
-
-Here’s a workaround:
-
-1.  Delete said color
-2.  Re-add the color (with the same name), this time specifying the new value
-    (following the steps described above in “Adding colors”)
-3.  **[Open a new issue][issue-template-color-updated]**
-
-When updating a color, please also remember to update its `filter` metadata. Filter values can be generated using [this CodePen](https://codepen.io/kaelig/full/jeObGP/) and are stored in [`/data/color-metadata.yml`](https://github.com/Shopify/polaris-tokens/blob/master/data/color-metadata.yml).
+Color names, values and descriptions are managed in [a Google Spreadsheet](https://docs.google.com/spreadsheets/d/1HzVgcxnnOCeB2zI46gcSd5fPFUJ2R03CSldx_SKQQwA/edit#gid=0).
 
 ### Developers
 
-This script fetches color data from the DSM API and dumps it in [`tokens/colors.yml`](https://github.com/Shopify/polaris-tokens/blob/master/tokens/colors.yml) under a format appropriate for consumption by [Theo](https://github.com/salesforce-ux/theo):
+This script fetches color data from the Google Spreadsheet and dumps it in [`tokens/colors.yml`](https://github.com/Shopify/polaris-tokens/blob/master/tokens/colors.yml) under a format appropriate for consumption by [Theo](https://github.com/salesforce-ux/theo):
 
 ```
-yarn get-tokens-from-invision-dsm
+yarn get-tokens-from-google-spreadsheets
 yarn dist
 yarn test -u
 ```
