+++ package.json
@@ -8,7 +8,7 @@
     "test": "sewing-kit test",
     "test-watch": "nodemon --exec 'npm-run-all dist test' --watch ./tokens --watch ./formats --watch ./gulpfile.js --watch ./scripts",
     "clean": "rimraf ./dist",
-    "get-tokens-from-invision-dsm": "node ./scripts/get-tokens-from-invision-dsm.js",
+    "get-tokens-from-google-spreadsheets": "node ./scripts/get-tokens-from-google-spreadsheets.js",
     "predist": "yarn clean",
     "dist": "npm-run-all dist:**",
     "dist-ci": "npm-run-all dist:* dist:*:*",
@@ -58,9 +58,7 @@
     "ase-util": "^1.0.1",
     "ase-utils": "^0.1.1",
     "browser-sync": "^2.26.3",
-    "dashify": "^2.0.0",
     "fs-extra": "^7.0.1",
-    "gray-matter": "^4.0.1",
     "gulp": "^3.9.1",
     "gulp-load-plugins": "^1.5.0",
     "gulp-plumber": "^1.2.1",
